+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-caller-service/src/main/java/com/tencent/cloud/polaris/discovery/service/caller/DiscoveryCallerService.java
@@ -27,7 +27,7 @@ import org.springframework.web.client.RestTemplate;
 /**
  * Discovery caller application.
  *
- * @author Haotian Zhang
+ * @author Haotian Zhang, cheese8
  */
 @SpringBootApplication
 @EnableFeignClients